backend_utilites/requirements.txt
pdfplumber
python-docx
sqlparse